requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Faker==16.3.0
Markdown==3.4.1
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.7
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
toposort==1.8
